Improve log messages for block values
